backend/requirements.txt
fastapi==0.110.1
uvicorn==0.25.0
boto3>=1.34.129
requests-oauthlib>=2.0.0
cryptography>=42.0.8
python-dotenv>=1.0.1
pymongo==4.5.0
pydantic>=2.6.4
email-validator>=2.2.0
pyjwt>=2.10.1
bcrypt==4.1.3
passlib>=1.7.4
tzdata>=2024.2
motor==3.3.1
pytest>=8.0.0
black>=24.1.1
isort>=5.13.2
flake8>=7.0.0
mypy>=1.8.0
python-jose>=3.3.0
requests>=2.31.0
pandas>=2.2.0
numpy>=1.26.0
python-multipart>=0.0.9
jq>=1.6.0
typer>=0.9.0
requests[security]>=2.31.0
pluginbase>=1.0.1
jinja2>=3.1.2
javalang==0.13.0
click>=8.1.7
cryptography>=42.0.0
setuptools>=69.0.3
typing-extensions>=4.10.0
xmltodict>=0.13.0
lxml>=5.1.0

click==8.3.0
cryptography==46.0.2
fastapi==0.110.1
javalang==0.13.0
lxml==6.0.2
motor==3.3.1
pluginbase==1.0.1
pydantic==2.12.0
pydantic_core==2.41.1
python-multipart==0.0.20
requests==2.32.5
requests-oauthlib==2.0.0
uvicorn==0.25.0
xmltodict==1.0.2
qark
